Move transform_constants method from Model to Parser
